# api/src/generated/models/job.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Job Rescale Job information
//
// swagger:model Job
type Job struct {

	// the container labels
	Commands []string `json:"commands"`

	// ended unix timestamp
	// Format: date-time
	Ended strfmt.DateTime `json:"ended,omitempty"`

	// A link to an external status page
	ExternalLink string `json:"external_link,omitempty"`

	// Job ID
	// Required: true
	ID *string `json:"id"`

	// the image ID
	Image string `json:"image,omitempty"`

	// the container labels
	Mounts []string `json:"mounts"`

	// platform
	// Enum: [kubernetes rescale]
	Platform string `json:"platform,omitempty"`

	// started unix timestamp
	// Format: date-time
	Started strfmt.DateTime `json:"started,omitempty"`

	// the status of the job
	Status string `json:"status,omitempty"`
}

// Validate validates this job
func (m *Job) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateEnded(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePlatform(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateStarted(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Job) validateEnded(formats strfmt.Registry) error {

	if swag.IsZero(m.Ended) { // not required
		return nil
	}

	if err := validate.FormatOf("ended", "body", "date-time", m.Ended.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *Job) validateID(formats strfmt.Registry) error {

	if err := validate.Required("id", "body", m.ID); err != nil {
		return err
	}

	return nil
}

var jobTypePlatformPropEnum []interface{}

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["kubernetes","rescale"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		jobTypePlatformPropEnum = append(jobTypePlatformPropEnum, v)
	}
}

const (

	// JobPlatformKubernetes captures enum value "kubernetes"
	JobPlatformKubernetes string = "kubernetes"

	// JobPlatformRescale captures enum value "rescale"
	JobPlatformRescale string = "rescale"
)

// prop value enum
func (m *Job) validatePlatformEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, jobTypePlatformPropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *Job) validatePlatform(formats strfmt.Registry) error {

	if swag.IsZero(m.Platform) { // not required
		return nil
	}

	// value enum
	if err := m.validatePlatformEnum("platform", "body", m.Platform); err != nil {
		return err
	}

	return nil
}

func (m *Job) validateStarted(formats strfmt.Registry) error {

	if swag.IsZero(m.Started) { // not required
		return nil
	}

	if err := validate.FormatOf("started", "body", "date-time", m.Started.String(), formats); err != nil {
		return err
	}

	return nil
}

// MarshalBinary interface implementation
func (m *Job) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Job) UnmarshalBinary(b []byte) error {
	var res Job
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}