# spec/openapi.yaml
swagger: '2.0'
info:
  title: AI Gateway
  version: 1.0.0
  description: |
    A platform for machine learning & high performance computing
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
host: 'localhost:9000'
basePath: /api/v1
securityDefinitions:
  api-authorizer:
    name: Authorization
    type: apiKey
    in: header
paths:
...
  /jobs:
    get:
      description: |
        returns training jobs on cloud
      operationId: getJobs
      tags:
        - job
      security:
        - api-authorizer: []
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Job'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      description: |
        Submit a job with the specified image
      operationId: postNewJob
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: body
          in: body
          required: true
          schema:
            title: JobAttrs
            type: object
            properties:
              platform_id:
                description: Platform ID
                type: string
                enum:
                    - kubernetes
                    - rescale
              notebook_id:
                description: Notebook container ID
                type: string
              entrypoint_file:
                description: The entrypoint file of the job
                type: string
              commands:
                description: commands to be excuted after the entrypoint
                type: array
                items:
                  type: string
              cpu:
                description: Requesting millicores of CPU
                type: integer
                format: int64
              mem:
                description: Requesting bytes of memory
                type: integer
                format: int64
              gpu:
                description: Requesting number of GPU
                type: integer
                format: int64
              coretype:
                description: Rescale CoreType as its infrastructure
                type: string
              cores:
                description: The number of CPU cores
                type: integer
                format: int64
      responses:
        '201':
          description: OK
          schema:
            title: newJobID
            type: object
            properties:
              id:
                description: ID of the new job
                type: string
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/jobs/{id}':
    get:
      description: |
        returns the details of a job
      operationId: getJobDetail
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Job ID
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/JobDetail'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    patch:
      description: |
        modify the job status
      operationId: modifyJob
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Job ID
          type: string
          required: true
        - name: body
          in: body
          required: true
          schema:
            title: JobAttrs
            type: object
            properties:
              status:
                description: status
                type: string
                enum:
                    - stopped
      responses:
        '200':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      description: |
        delete a job
      operationId: deleteJob
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Job ID
          type: string
          required: true
      responses:
        '204':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/jobs/{id}/logs':
    get:
      description: |
        returns the logs of a job
      operationId: getJobLogs
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Job ID
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/JobLogs'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/jobs/{id}/files':
    get:
      description: |
        returns the list of output files
      operationId: getJobFiles
      tags:
        - job
      security:
        - api-authorizer: []
      parameters:
        - name: id
          in: path
          description: Job ID
          type: string
          required: true
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/JobFiles'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Job:
    description: Rescale Job information
    required:
      - id
    properties:
      id:
        description: Job ID
        type: string
      platform:
        description: platform
        type: string
        enum:
          - kubernetes
          - rescale
      status:
        description: the status of the job
        type: string
      image:
        description: the image ID
        type: string
      mounts:
        description: the container labels
        type: array
        items:
          type: string
      commands:
        description: the container labels
        type: array
        items:
          type: string
      started:
        description: started unix timestamp
        type: string
        format: date-time
      ended:
        description: ended unix timestamp
        type: string
        format: date-time
      external_link:
        description: A link to an external status page
        type: string
  JobDetail:
    description: the details of a job
    allOf: 
      - $ref: "#/definitions/Job"
      - $ref: "#/definitions/JobLogs"
      - $ref: "#/definitions/JobFiles"
  JobLog:
    description: the log structure
    required:
      - log
    properties:
      time:
        type: string
        format: date-time
      log:
        type: string
  JobLogs:
    description: the log structure
    required:
      - logs
    properties:
      logs:
        type: array
        items:
          $ref: '#/definitions/JobLog'
  JobFile:
    description: the output file
    required:
      - name
      - size
    properties:
      name:
        type: string
      size:
        type: integer
        format: int64
      downloadURL:
        type: string
  JobFiles:
    description: the output files
    required:
      - files
    properties:
      files:
        type: array
        items:
          $ref: '#/definitions/JobFile'
      apiToken:
        type: string
...
  Error:
    required:
      - code
      - message
    properties:
      code:
        type: string
      message:
        type: string